Move ToVariant into variant.rs.
